[#1] AcornDemoTests TestCode, TestHelper 추가, AcornDemoUITest 삭제

+++ Acorn/Acorn/Sources/DiskCache.swift
@@ -11,8 +11,8 @@ public class DiskCache {
     private var fileDirectoryURL: URL? {
         return FileManager.default.urls(for: .cachesDirectory, in: .userDomainMask).first?.appendingPathComponent("cache")
     }
-    private var maximumDiskBytes: Int
     private var expiration: CacheExpiration
+    private(set) var maximumDiskBytes: Int
 
     init(maximumDiskBytes: Int, expiration: CacheExpiration) {
         self.maximumDiskBytes = maximumDiskBytes
@@ -28,10 +28,12 @@ public extension DiskCache {
             debugPrint(ImageCacheError.invalidFileURL.description)
             return nil
         }
+        
         guard let imageData = try? Data(contentsOf: fileURL) else {
             debugPrint(ImageCacheError.failedReadDataFromDisk.description)
             return nil
         }
+        
         return CachedImage(imageData: imageData)
     }
 
@@ -42,57 +44,74 @@ public extension DiskCache {
         }
         do {
             try data.imageData.write(to: fileURL)
-        } catch {
-            debugPrint(ImageCacheError.failedSaveDataToDisk.description)
-            return
-        }
-        
-        let now = Date()
-        let attributes: [FileAttributeKey: Any] = [
-            .creationDate: now.fileDate,
-            .modificationDate: expiration.calculateExpirationDate(from: now).fileDate
-        ]
-        do {
+            
+            let now = Date()
+            let attributes: [FileAttributeKey: Any] = [
+                .creationDate: now.fileDate,
+                .modificationDate: expiration.calculateExpirationDate(from: now).fileDate
+            ]
             try FileManager.default.setAttributes(attributes, ofItemAtPath: fileURL.path)
-        } catch {
-            debugPrint(ImageCacheError.failedSetCacheFileAttribute.description)
-            return
-        }
-    }
-    
-    func removeOldCache() {
-        let urlResourceKeys: [[URLResourceKey]] = [[
-                                                        .isDirectoryKey,
-                                                        .contentModificationDateKey,
-                                                   ],
-                                                   [
-                                                        .isDirectoryKey,
-                                                        .creationDateKey,
-                                                        .fileSizeKey
-                                                   ]]
-        do {
-            for i in 0..<2 {
-                let urls = try getAllFileURLs(key: urlResourceKeys[i])
-                let keys = Set(urlResourceKeys[i])
-                if i == 0 {
-                    removeExpiredCache(urls: urls, keys: keys)
-                } else {
-                    try removeOverSizeCache(urls: urls, keys: keys)
-                }
-            }
-        } catch ImageCacheError.invalidFileDirectoryURL {
-            debugPrint(ImageCacheError.invalidFileDirectoryURL.description)
-        } catch ImageCacheError.failedCreateFileEnumerator {
-            debugPrint(ImageCacheError.failedCreateFileEnumerator.description)
-        } catch ImageCacheError.invalidFileEnumeratorContents {
-            debugPrint(ImageCacheError.invalidFileEnumeratorContents.description)
-        } catch ImageCacheError.failedSortedArray {
-            debugPrint(ImageCacheError.failedSortedArray.description)
         } catch {
             debugPrint(error.localizedDescription)
         }
     }
-    
+
+    func getTotalDiskCacheSize() -> Int {
+        let key: [URLResourceKey] = [.fileSizeKey]
+        guard let urls = try? getAllFileURLs(key: key) else {
+            return 0
+        }
+        
+        let totalSize = urls.reduce(0) { size, url in
+            let resourceValues = try? url.resourceValues(forKeys: Set(key))
+            let fileSize = resourceValues?.fileSize ?? 0
+            return size + fileSize
+        }
+        return totalSize
+    }
+
+    func removeUnnecessaryCache() {
+        do {
+            try removeExpiredCache()
+            try removeOverSizeCache()
+        } catch {
+            if let error = error as? ImageCacheError {
+                debugPrint(error.description)
+            } else {
+                debugPrint(error.localizedDescription)
+            }
+        }
+    }
+
+    func removeExpiredCache() throws {
+        let urlResourceKeys: [URLResourceKey] = [.isDirectoryKey,
+                                                 .contentModificationDateKey]
+        do {
+            let urls = try getAllFileURLs(key: urlResourceKeys)
+            let keys = Set(urlResourceKeys)
+            let expiredFiles = getExpiredFiles(urls: urls, keys: keys)
+
+            expiredFiles.forEach { url in
+                removeFile(at: url)
+            }
+        } catch {
+            throw error
+        }
+    }
+
+    func removeOverSizeCache() throws {
+        let urlResourceKeys: [URLResourceKey] = [.isDirectoryKey,
+                                                 .creationDateKey,
+                                                 .fileSizeKey]
+        do {
+            let urls = try getAllFileURLs(key: urlResourceKeys)
+            let keys = Set(urlResourceKeys)
+            try removeDiskCacheUntilTargetSize(urls: urls, keys: keys)
+        } catch {
+            throw error
+        }
+    }
+
     func clearDiskCache() {
         if let fileDirectoryURL = self.fileDirectoryURL {
             removeFile(at: fileDirectoryURL)
@@ -144,49 +163,23 @@ private extension DiskCache {
         return urls
     }
 
-    func removeExpiredCache(urls: [URL], keys: Set<URLResourceKey>) {
-        let expiredFiles = getExpiredFiles(urls: urls, keys: keys)
-
-        expiredFiles.forEach { url in
-            removeFile(at: url)
-        }
-    }
-
     func getExpiredFiles(urls: [URL], keys: Set<URLResourceKey>) -> [URL] {
         let now = Date()
 
         let expiredURLs = urls.filter { fileURL in
             let resourceValues = try? fileURL.resourceValues(forKeys: keys)
-            let exprirationDate = resourceValues?.contentModificationDate
+            let expirationDate = resourceValues?.contentModificationDate
             let isDirectory = resourceValues?.isDirectory ?? false
             if isDirectory {
                 return false
             }
-            return exprirationDate?.isPast(from: now) ?? true
+            return expirationDate?.isPast(from: now) ?? true
         }
         return expiredURLs
     }
 
-    func removeOverSizeCache(urls: [URL], keys: Set<URLResourceKey>) throws {
-        var totalSize = getTotalDiskCacheSize(urls: urls, keys: keys)
-        let targetSize = maximumDiskBytes / 2
-        guard var sortURLsByCreationDate = try sortURLsByCreationDate(urls: urls, keys: keys) else {
-            throw ImageCacheError.failedSortedArray
-        }
-        removeDiskCacheUntilTargetSize(totalSize: &totalSize, targetSize: targetSize, sortURLsByCreationDate: &sortURLsByCreationDate, keys: keys)
-    }
-
-    func getTotalDiskCacheSize(urls: [URL], keys: Set<URLResourceKey>) -> Int {
-        let totalSize = urls.reduce(0) { size, url in
-            let resourceValues = try? url.resourceValues(forKeys: keys)
-            let fileSize = resourceValues?.fileSize ?? 0
-            return size + fileSize
-        }
-        return totalSize
-    }
-
-    func sortURLsByCreationDate(urls: [URL], keys: Set<URLResourceKey>) throws -> [URL]? {
-        let sortedUrls = try? urls.sorted(by: {
+    func sortURLsByCreationDate(urls: [URL], keys: Set<URLResourceKey>) throws -> [URL] {
+        let sortedUrls = try urls.sorted(by: {
             let resourceValuesForLhs = try $0.resourceValues(forKeys: keys)
             let resourceValuesForRhs = try $1.resourceValues(forKeys: keys)
             guard let creationDateForLhs = resourceValuesForLhs.creationDate,
@@ -198,12 +191,21 @@ private extension DiskCache {
         return sortedUrls
     }
 
-    func removeDiskCacheUntilTargetSize(totalSize: inout Int, targetSize: Int, sortURLsByCreationDate: inout [URL], keys: Set<URLResourceKey>) {
-        while totalSize > targetSize, let lastURL = sortURLsByCreationDate.popLast() {
-            let resourceValues = try? lastURL.resourceValues(forKeys: keys)
-            let fileSize = resourceValues?.fileSize ?? 0
-            totalSize -= fileSize
-            removeFile(at: lastURL)
+    func removeDiskCacheUntilTargetSize(urls: [URL], keys: Set<URLResourceKey>) throws {
+        do {
+            let targetSize = maximumDiskBytes / 2
+            var totalSize = getTotalDiskCacheSize()
+            var sortedURLsByCreationDate = try sortURLsByCreationDate(urls: urls, keys: keys)
+            
+            while totalSize > targetSize, let lastURL = sortedURLsByCreationDate.popLast() {
+                let resourceValues = try lastURL.resourceValues(forKeys: keys)
+                if let fileSize = resourceValues.fileSize {
+                    totalSize -= fileSize
+                    removeFile(at: lastURL)
+                }
+            }
+        } catch {
+            throw error
         }
     }
